% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tic.R
\name{calculate_iic}
\alias{calculate_iic}
\title{calculate item information.}
\usage{
calculate_iic(object, item, theta, breaks, grm, use_ltm)
}
\arguments{
\item{object}{mirt or grm object. If you use the grm object, you must set use_ltm = TRUE.}

\item{item}{numeric. The number of the item to be visualized. The length must be 1.}

\item{theta}{vector. The length must be 2. Specify the range of theta in the IRCCC to be visualized using c().}

\item{breaks}{numeric. Specifies the number of divisions of theta. For example, if you specify 1000, the range specified by theta will be divided into 1000 parts. The default is 100.}

\item{grm}{logical. The default is TRUE. The default is TRUE. If you are using a binary IRT (2PLM) object, set it to TRUE.}

\item{use_ltm}{logical. The default is FALSE. If you are using an object (either an ltm object or a grm object) from the ltm package, set this to TRUE. When use_ltm = TRUE, the scale factor D is corrected to 1.701.}
}
\description{
\code{calculate_iic} Calculate the item information volume from the item parameter obtained from GRM or 2PLM.
}
\keyword{internal}
