% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_irccc.R
\name{get_par}
\alias{get_par}
\title{Extracts item parameters from the mirt object.}
\usage{
get_par(object)
}
\arguments{
\item{object}{mirt object}
}
\description{
\code{get_par} Extracts item parameters from the mirt object.
}
\keyword{internal}
