% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tic.R
\name{gg_iic}
\alias{gg_iic}
\title{visualize Item Information Curves.}
\usage{
gg_iic(
  object,
  item,
  theta = c(-3, 3),
  breaks = 100,
  grm = TRUE,
  use_ltm = FALSE,
  output.data = FALSE
)
}
\arguments{
\item{object}{mirt or grm object. If you use the grm object, you must set ltm = TRUE.}

\item{item}{numeric. The number of the item to be visualized. The length must be 1.}

\item{theta}{vector. The length must be 2. Specify the range of θ in the IRCCC to be visualized using c().}

\item{breaks}{numeric. Specifies the number of divisions of theta. For example, if you specify 1000, the range specified by theta will be divided into 1000 parts. The default is 100.}

\item{grm}{logical. The default is TRUE. If you are using a binary IRT　(2PLM) object, set it to TRUE.}

\item{use_ltm}{logical. The default is FALSE. If you are using an object (either an ltm object or a grm object) from the ltm package, set this to TRUE.}

\item{output.data}{Logical. The default is FALSE. If TRUE, instead of visualizing IRCCC, the data that forms the basis of IRCCC is output in long format.}
}
\value{
Visualize item information using ggplot2 package.
}
\description{
\code{gg_iic} Calculate the item information volume from the item parameter obtained from GRM or 2PLM.
}
\examples{
#library(mirt)
#mirt_object <- mirt(data, itemtype = "graded")

#gg_iic(mirt_object, item = 1)
#gg <- gg_iic(mirt_object, item = 1)
#gg + theme_bw()

#library(ltm)
#ltm_object <- ltm(data ~ z1, IRT.param = TRUE)

#gg_iic(ltm_object, item = 1, theta = c(-5, 5), breaks = 1000, grm = FALSE, ltm = TRUE)
}
