% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_irccc.R
\name{gg_tic}
\alias{gg_tic}
\title{visualize Test Information Curve from mirt object.}
\usage{
gg_tic(object, theta = c(-3, 3), breaks = 100, se = TRUE)
}
\arguments{
\item{object}{mirt object.}

\item{theta}{vector. The length must be 2. Specify the range of θ in the  TIC to be visualized using c().}

\item{breaks}{numeric. Specifies the number of divisions of theta. For example, if you specify 1000, the range specified by theta will be divided into 1000 parts. The default is 100.}

\item{se}{logical. The default is TRUE. Adds the standard error of the test information to the TIC. If FALSE, only the TIC is visualized.}
}
\value{
This function visualizes TIC using ggplot2.
}
\description{
\code{gg_irccc} visualize TIC from mirt object.
}
\examples{
#gg_tic(object = res)
#plot1 <- gg_tic(object = res)
#plot1 + theme_bw()

#gg_tic(object = res, se = FALSE)
}
