% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_irccc.R
\name{calculate_irccc}
\alias{calculate_irccc}
\title{A category response probability is calculated from the item parameter.}
\usage{
calculate_irccc(thetas, a, b, item, k, use_ltm)
}
\arguments{
\item{thetas}{vector. Theta is the latent value of the range for calculating the category response probability. This value is calculated from the theta and breaks of gg_irccc.}

\item{a}{numeric. The length must be 1. This represents the identification parameter.}

\item{b}{vector. This is the parameter for difficulty level.}

\item{item}{numeric. The length must be 1. This value is the number of the item for calculating the category response probability.}

\item{k}{numeric. Number of rating categories for any given item. It must be 3 or more.}

\item{use_ltm}{logical. The default is FALSE. If you are using the mirt object, set this to FALSE. If you are using the grm object from the ltm package, set this option to TRUE.}
}
\description{
\code{calculate_irccc} Calculate a category response probability from the item parameter.
}
\keyword{internal}
