% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_tic.R
\name{gg_tic}
\alias{gg_tic}
\title{visualize Item Information Curves.}
\usage{
gg_tic(
  object,
  theta = c(-3, 3),
  breaks = 100,
  se = TRUE,
  grm = TRUE,
  use_ltm = FALSE,
  output.data = FALSE
)
}
\arguments{
\item{object}{mirt or grm object. If you use the grm object, you must set ltm = TRUE.}

\item{theta}{vector. The length must be 2. Specify the range of theta in the IRCCC to be visualized using c().}

\item{breaks}{numeric. Specifies the number of divisions of theta. For example, if you specify 1000, the range specified by theta will be divided into 1000 parts. The default is 100.}

\item{se}{logical. The default is TRUE. If you are not adding a SE to the test information curve, set it to FALSE.}

\item{grm}{logical. The default is TRUE. If you are using a binary IRT (2PLM) object, set it to TRUE.}

\item{use_ltm}{logical. The default is FALSE. If you are using an object (either an ltm object or a grm object) from the ltm package, set this to TRUE.When use_ltm = TRUE, the scale factor D is corrected to 1.701.}

\item{output.data}{logical. The default is FALSE. If TRUE, instead of visualizing IRCCC, the data that forms the basis of IRCCC is output in long format.}
}
\value{
Visualize test information Curves using ggplot2 package.
}
\description{
\code{gg_tic} Calculate the item information volume from the item parameter obtained from GRM.
}
\examples{
#library(mirt)
#mirt_object <- mirt(data, model = 1, itemtype = "graded")

#gg_iic(mirt_object, item = 1)
#gg <- gg_iic(mirt_object, item = 1)
#gg + theme_bw()

#library(ltm)
#ltm_object <- ltm(data ~ z1, IRT.param = TRUE)

#gg_iic(ltm_object, item = 1, theta = c(-5, 5), breaks = 1000, grm = FALSE, use_ltm = TRUE)
}
